Create randomsort.m
Use randomsort(x,1) to execute in Matlab. X acts as the size of the initial randomized array. It then creates a plot to show the randomized array, and shows it sort said array. It can be reset from figure GUI, and the size of the array can be changed from there.